refactor engine
