+++ internal/types.go
@@ -1,0 +1,31 @@
+package internal
+
+import (
+	"go/token"
+	"os"
+	"strings"
+)
+
+// SourceCode stores the content of a source code file.
+type SourceCode struct {
+	Lines []string
+}
+
+// ReadSourceFile reads the content of a file and returns it as a `SourceCode` struct.
+func ReadSourceCode(filename string) (*SourceCode, error) {
+	content, err := os.ReadFile(filename)
+	if err != nil {
+		return nil, err
+	}
+	lines := strings.Split(string(content), "\n")
+	return &SourceCode{Lines: lines}, nil
+}
+
+// Issue represents a lint issue found in the code base.
+type Issue struct {
+	Rule     string
+	Filename string
+	Start    token.Position
+	End      token.Position
+	Message  string
+}
